+++ app/Http/Controllers/TaskController.php
@@ -9,7 +9,7 @@ use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
-
+use Illuminate\Support\Facades\Log;
 class TaskController extends Controller
 {
     protected $task;
@@ -34,7 +34,8 @@ class TaskController extends Controller
 
     public function store(TaskRequest $request)
     {
-        $this->task->create($request->all());
+        Task::create($request->validated());
+    
         return response()->json(['success' => true]);
     }
 
@@ -46,10 +47,8 @@ class TaskController extends Controller
 
     public function update(TaskRequest $request, Task $task)
     {
-
-        $this->$task->update($request->all());
-
-        return response()->json(['success' => true]);
+        $task->update($request->validated());
+        return response()->json(['message' => 'Task updated successfully']);
     }
 
     public function destroy(Task $task)
@@ -66,52 +65,34 @@ class TaskController extends Controller
     public function search(Request $request)
     {
         try {
-            $query = $request->input('query', '');
-            $projectId = $request->input('project_id', '');
-            $page = $request->input('page', 1);
-            $sortColumn = $request->get('sort_column', 'id');
-            $sortOrder = $request->get('sort_order', 'desc');
-
-            $allowedSortColumns = ['id', 'name', 'content'];
-            if (!in_array($sortColumn, $allowedSortColumns)) {
-                $sortColumn = 'id';
-            }
-
-            $allowedSortOrders = ['asc', 'desc'];
-            if (!in_array(strtolower($sortOrder), $allowedSortOrders)) {
-                $sortOrder = 'desc';
-            }
-
-            $tasksQuery = Task::query()
-                ->when($query, function ($q) use ($query) {
-                    return $q->where('name', 'like', "%{$query}%");
-                })
-                ->when($projectId, function ($q) use ($projectId) {
-                    return $q->where('project_id', $projectId);
-                })
-                ->with('project')
-                ->orderBy($sortColumn, $sortOrder);
-
-            if (!empty($projectId)) {
-                $tasksQuery->where('project_id', $projectId);
-            }
-
-            $tasks = $tasksQuery->orderBy($sortColumn, $sortOrder)
-                ->latest('id')
-                ->paginate(10, ['*'], 'page', $page);
-
+            $tasks = $this->buildTaskQuery($request)->paginate(10, ['*'], 'page', $request->input('page', 1));
             return response()->json([
                 'data' => $tasks->items(),
-                'links' => $tasks->appends([
-                    'query' => $query,
-                    'project_id' => $projectId,
-                    'sort_column' => $sortColumn,
-                    'sort_order' => $sortOrder
-                ])->links()->toHtml()
+                'links' => $tasks->appends($request->all())->links()->toHtml(),
             ]);
         } catch (\Exception $e) {
-            \Log::error('Search error: ' . $e->getMessage());
+            Log::error('Search error: ' . $e->getMessage());
             return response()->json(['error' => 'An error occurred'], 500);
         }
+    }
+
+    protected function buildTaskQuery(Request $request)
+    {
+        $allowedSortColumns = ['id', 'name', 'content'];
+        $sortColumn = in_array($request->get('sort_column', 'id'), $allowedSortColumns) ? $request->get('sort_column') : 'id';
+
+        $allowedSortOrders = ['asc', 'desc'];
+        $sortOrder = in_array(strtolower($request->get('sort_order', 'desc')), $allowedSortOrders) ? $request->get('sort_order') : 'desc';
+
+        return Task::query()
+            ->when($request->input('query'), function ($q, $query) {
+                return $q->where('name', 'like', "%{$query}%");
+            })
+            ->when($request->input('project_id'), function ($q, $projectId) {
+                return $q->where('project_id', $projectId);
+            })
+            ->with('project')
+            ->orderBy($sortColumn, $sortOrder)
+            ->latest('id');
     }
 }
