+++ app/Models/Task.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Str;
 
 class Task extends Model
 {
@@ -17,4 +18,12 @@ class Task extends Model
     {
         return $this->belongsTo(Project::class);
     }
+    public function getContentAttribute($value)
+    {
+        return Str::limit($value, 100);
+    }
+    // public function getNameAttribute($value)
+    // {
+    //     return Str::limit($value, 20);
+    // }
 }
