+++ resources/views/layouts/app.blade.php
@@ -21,7 +21,7 @@
     <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.3/dist/umd/popper.min.js"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-    
+
 
     <!-- href="{{ asset('') }}" -->
     <!-- test  -->
@@ -42,12 +42,11 @@
     <link href="{{ asset('user/vendor/quill/quill.bubble.css') }}" rel="stylesheet">
     <link href="{{ asset('user/vendor/remixicon/remixicon.css') }}" rel="stylesheet">
     <link href="{{ asset('user/vendor/simple-datatables/style.css') }}" rel="stylesheet">
-
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
 
     <!-- Template Main CSS File -->
     <link href="{{ asset('user/css/style.css') }}" rel="stylesheet">
     <link href="{{ asset('user/css/style-user.css') }}" rel="stylesheet">
-
 
 
     <!-- Scripts -->
@@ -62,7 +61,7 @@
             <div class="d-flex align-items-center justify-content-between">
                 <a href="{{ url('/home') }}" class="logo d-flex align-items-center">
 
-                    <span class="d-none d-lg-block">Delivery Management</span>
+                    <span class="d-none d-lg-block">Task Management</span>
                 </a>
                 <i class="bi bi-list toggle-sidebar-btn"></i>
             </div>
@@ -95,15 +94,14 @@
                                     data-bs-toggle="dropdown">
                                     <img src="{{ asset('user/img/profile-img.jpg') }}" alt="Profile"
                                         class="rounded-circle">
-                                    &nbsp; <span class="fs-5">{{ Auth::user()->role }}</span> 
+                                    &nbsp; <span class="fs-5">{{ Auth::user()->role }}</span>
                                     {{ Auth::user()->name }}
                                 </a><!-- End Profile Iamge Icon -->
 
                                 <ul class="dropdown-menu dropdown-menu-end dropdown-menu-arrow profile">
                                     <li class="dropdown-header">
-                                        <h5><a class="nav-link " href="#" role="button"
-                                                data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"
-                                                v-pre>
+                                        <h5><a class="nav-link " href="#" role="button" data-bs-toggle="dropdown"
+                                                aria-haspopup="true" aria-expanded="false" v-pre>
                                                 <span class="fs-5">{{ Auth::user()->role }}</span>
                                                 {{ Auth::user()->name }}
                                             </a></h5>
@@ -147,8 +145,7 @@
                                     </li> --}}
 
                                     <li>
-                                        <div class="dropdown-item align-items-center"
-                                            aria-labelledby="navbarDropdown">
+                                        <div class="dropdown-item align-items-center" aria-labelledby="navbarDropdown">
 
                                             <a class="dropdown-item" href="{{ route('logout') }}"
                                                 onclick="event.preventDefault();
@@ -184,9 +181,9 @@
                     </a>
                 </li><!-- End Dashboard Nav -->
 
-                
 
-                
+
+
                 <!-- End Tables Nav -->
 
                 <!-- <li class="nav-item">
