+++ resources/views/tasks/index.blade.php
@@ -7,19 +7,24 @@
                 <div class="col-md-6">
                     <button class="btn btn-primary" id="add-task-btn">Add Task</button>
                 </div>
-                <div class="col-md-6 d-flex justify-content-end align-items-center">
-                    <form id="filter-form" class="form-inline">
-                        <div class="input-group">
-                            <select id="project-filter" class="form-control ml-2">
-                                <option value="">All Projects</option>
-                                @foreach ($projects as $project)
-                                    <option value="{{ $project->id }}">{{ $project->name }}</option>
-                                @endforeach
-                            </select>
-                            <input id="search-input" type="text" class="form-control" placeholder="Search tasks">
-                        </div>
-                    </form>
-                </div>
+                <div class="col-md-6 d-flex justify-content-end align-items-center">  
+                    <form id="filter-form" class="form-inline">  
+                        <div class="input-group">  
+                            
+                            <div class="position-relative">  
+                                <select id="project-filter" class="form-control ml-2">  
+                                    <option value="">All Projects</option>  
+                                    @foreach ($projects as $project)  
+                                        <option value="{{ $project->id }}">{{ $project->name }}</option>  
+                                    @endforeach  
+                                </select>  
+                                <i class="bi bi-chevron-down position-absolute" style="right: 10px; top: 50%; transform: translateY(-50%); pointer-events: none;"></i>  
+                            </div>  
+                            
+                            <input id="search-input" type="text" class="form-control" placeholder="Search tasks">  
+                        </div>  
+                    </form>  
+                </div>  
             </div>
             <div class="row justify-content-center">
                 <table class="table table-striped">
@@ -64,8 +69,11 @@
                                 <td style="height: 80px;width: 50%">{{ $task->content }}</td>
                                 <td>{{ $task->project ? $task->project->name : 'No Project' }}</td>
                                 <td class="text-center">
-                                    <button class="btn btn-danger delete-btn" data-task-id="{{ $task->id }}">Delete</button>
-                                    <button class="btn btn-warning edit-btn" data-task-id="{{ $task->id }}" style="display:inline-block;margin-left: 3px">Edit</button>
+                                    <button class="btn btn-warning edit-btn" data-task-id="{{ $task->id }}"
+                                        style="display:inline-block;margin-left: 3px"><i class="bi bi-pencil"></i></button>
+                                    <button class="btn btn-danger delete-btn"
+                                        data-task-id="{{ $task->id }}"><i class="bi bi-trash"></i></button>
+                                    
                                 </td>
                             </tr>
                         @endforeach
@@ -83,23 +91,8 @@
     </div>
 
     <!-- Modal Confirm Delete -->
-    <div class="modal fade" id="deleteConfirmModal" tabindex="-1" aria-labelledby="deleteConfirmModalLabel" aria-hidden="true">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="deleteConfirmModalLabel">Confirm Delete</h5>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                </div>
-                <div class="modal-body">
-                    Are you sure you want to delete this task?
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-                    <button type="button" class="btn btn-danger" id="confirm-delete-btn">Delete</button>
-                </div>
-            </div>
-        </div>
-    </div>
+    @include('tasks.delete-modal')
+
 
     <!-- Modal Create Task -->
     @include('tasks.create-modal')
@@ -107,275 +100,9 @@
     <!-- Modal Edit Task -->
     @include('tasks.edit-modal')
 
-    <style>
-        .sort-icon {
-            font-size: 1em;
-            margin-left: 5px;
-            display: inline-block;
-            transition: opacity 0.3s;
-        }
+    <!-- Script tasks -->
+    <script src="assets{{'/js/tasks.js'}}"></script>
 
-        .sort-asc .bi-sort-alpha-up {
-            color: #000;
-        }
 
-        .sort-desc .bi-sort-alpha-down {
-            color: #000;
-        }
-
-        .sort-link.active .sort-icon {
-            opacity: 1;
-        }
-    </style>
-
-    <script>
-        document.addEventListener('DOMContentLoaded', function() {
-            const searchInput = document.getElementById('search-input');
-            const projectFilter = document.getElementById('project-filter');
-            const tasksTableBody = document.getElementById('tasks-table-body');
-            const paginationLinks = document.getElementById('pagination-links');
-            const sortLinks = document.querySelectorAll('.sort-link');
-            const csrfToken = '{{ csrf_token() }}';
-
-            let currentPage = 1;
-            let currentQuery = '';
-            let currentProjectId = '';
-            let sortColumn = 'id';
-            let sortOrder = 'desc';
-            let taskToDelete = null;
-
-            // Hiển thị modal tạo mới nhiệm vụ
-            document.getElementById('add-task-btn').addEventListener('click', function() {
-                const modal = new bootstrap.Modal(document.getElementById('createTaskModal'));
-                modal.show();
-            });
-
-            // Hiển thị modal chỉnh sửa nhiệm vụ
-            document.addEventListener('click', function(event) {
-                if (event.target && event.target.classList.contains('edit-btn')) {
-                    const taskId = event.target.dataset.taskId;
-                    fetch(`/tasks/${taskId}/edit`)
-                        .then(response => response.json())
-                        .then(data => {
-                            const task = data.task;
-                            document.getElementById('edit-task-name').value = task.name;
-                            document.getElementById('edit-task-content').value = task.content;
-                            document.getElementById('edit-task-project').value = task.project_id || '';
-                            document.getElementById('edit-task-id').value = task.id;
-                            const modal = new bootstrap.Modal(document.getElementById('editTaskModal'));
-                            modal.show();
-                        })
-                        .catch(error => {
-                            console.error('Error fetching task:', error);
-                        });
-                }
-
-                if (event.target && event.target.classList.contains('delete-btn')) {
-                    const taskId = event.target.dataset.taskId;
-                    taskToDelete = taskId;
-                    const taskName = event.target.closest('tr').querySelector('td:nth-child(2)').textContent;
-                    const modalTitle = document.querySelector('#deleteConfirmModalLabel');
-                    const modalBody = document.querySelector('.modal-body');
-                    modalTitle.textContent = `Delete Task #${taskId}`;
-                    modalBody.textContent = `Are you sure you want to delete the task "${taskName}"?`;
-                    const modal = new bootstrap.Modal(document.getElementById('deleteConfirmModal'));
-                    modal.show();
-                }
-            });
-
-            // Xác nhận xóa nhiệm vụ
-            document.getElementById('confirm-delete-btn').addEventListener('click', function() {
-                if (taskToDelete) {
-                    fetch(`/tasks/${taskToDelete}`, {
-                            method: 'DELETE',
-                            headers: {
-                                'X-Requested-With': 'XMLHttpRequest',
-                                'X-CSRF-TOKEN': csrfToken
-                            }
-                        })
-                        .then(response => response.json())
-                        .then(() => {
-                            document.querySelector(`tr[data-task-id="${taskToDelete}"]`).remove();
-                            updateTable(currentQuery, currentProjectId, currentPage, sortColumn, sortOrder);
-                        })
-                        .catch(error => {
-                            console.error('Error deleting task:', error);
-                        })
-                        .finally(() => {
-                            taskToDelete = null;
-                            const modal = bootstrap.Modal.getInstance(document.getElementById('deleteConfirmModal'));
-                            if (modal) {
-                                modal.hide();
-                            }
-                        });
-                }
-            });
-
-            // Tạo nhiệm vụ mới
-            document.getElementById('create-task-form').addEventListener('submit', function(event) {
-                event.preventDefault();
-                const name = document.getElementById('task-name').value.trim();
-                const content = document.getElementById('task-content').value.trim();
-                const projectId = document.getElementById('task-project').value;
-                fetch('/tasks', {
-                        method: 'POST',
-                        headers: {
-                            'X-Requested-With': 'XMLHttpRequest',
-                            'X-CSRF-TOKEN': csrfToken,
-                            'Content-Type': 'application/json'
-                        },
-                        body: JSON.stringify({
-                            name: name,
-                            content: content,
-                            project_id: projectId
-                        })
-                    })
-                    .then(response => response.json())
-                    .then(() => {
-                        const modal = bootstrap.Modal.getInstance(document.getElementById('createTaskModal'));
-                        modal.hide();
-                        updateTable(currentQuery, currentProjectId, currentPage, sortColumn, sortOrder);
-                    })
-                    .catch(error => {
-                        console.error('Error creating task:', error);
-                    });
-            });
-
-            // Chỉnh sửa nhiệm vụ
-            document.getElementById('edit-task-form').addEventListener('submit', function(event) {
-                event.preventDefault();
-                const name = document.getElementById('edit-task-name').value.trim();
-                const content = document.getElementById('edit-task-content').value.trim();
-                const projectId = document.getElementById('edit-task-project').value;
-                const taskId = document.getElementById('edit-task-id').value;
-                fetch(`/tasks/${taskId}`, {
-                        method: 'PUT',
-                        headers: {
-                            'X-Requested-With': 'XMLHttpRequest',
-                            'X-CSRF-TOKEN': csrfToken,
-                            'Content-Type': 'application/json'
-                        },
-                        body: JSON.stringify({
-                            name: name,
-                            content: content,
-                            project_id: projectId
-                        })
-                    })
-                    .then(response => response.json())
-                    .then(() => {
-                        const modal = bootstrap.Modal.getInstance(document.getElementById('editTaskModal'));
-                        modal.hide();
-                        updateTable(currentQuery, currentProjectId, currentPage, sortColumn, sortOrder);
-                    })
-                    .catch(error => {
-                        console.error('Error updating task:', error);
-                    });
-            });
-
-            function updateTable(query, projectId, page = 1, column = 'id', order = 'desc') {
-                query = query.trim();
-                currentQuery = query;
-                currentPage = page;
-                currentProjectId = projectId;
-                sortColumn = column;
-                sortOrder = order;
-
-                const url = new URL('/api/tasks/search', window.location.origin);
-                url.searchParams.append('query', query);
-                url.searchParams.append('project_id', projectId);
-                url.searchParams.append('page', page);
-                url.searchParams.append('sort_column', column);
-                url.searchParams.append('sort_order', order);
-
-                fetch(url.toString(), {
-                        headers: {
-                            'X-Requested-With': 'XMLHttpRequest'
-                        }
-                    })
-                    .then(response => response.json())
-                    .then(data => {
-                        tasksTableBody.innerHTML = '';
-                        if (data.data && Array.isArray(data.data)) {
-                            data.data.forEach(task => {
-                                const row = document.createElement('tr');
-                                row.dataset.taskId = task.id;
-                                row.innerHTML = `
-                            <td>${task.id}</td>
-                            <td>${task.name}</td>
-                            <td>${task.content}</td>
-                            <td>${task.project ? task.project.name : ''}</td>
-                            <td>
-                                <button class="btn btn-warning edit-btn" data-task-id="${task.id}">Edit</button>
-                                <button class="btn btn-danger delete-btn" data-task-id="${task.id}">Delete</button>
-                            </td>
-                        `;
-                                tasksTableBody.appendChild(row);
-                            });
-                        }
-
-                        // Cập nhật phân trang
-                        if (data.links) {
-                            paginationLinks.innerHTML = data.links;
-
-                            paginationLinks.querySelectorAll('a').forEach(link => {
-                                link.addEventListener('click', function(event) {
-                                    event.preventDefault();
-                                    const url = new URL(link.href);
-                                    const page = url.searchParams.get('page');
-                                    updateTable(currentQuery, currentProjectId, page, sortColumn, sortOrder);
-                                });
-                            });
-                        } else {
-                            paginationLinks.innerHTML = '';
-                        }
-
-                        // Cập nhật các liên kết sắp xếp
-                        sortLinks.forEach(link => {
-                            const column = link.dataset.column;
-                            const order = link.dataset.order;
-                            link.classList.remove('sort-asc', 'sort-desc');
-                            if (column === sortColumn) {
-                                link.classList.add(order === 'asc' ? 'sort-asc' : 'sort-desc');
-                                link.dataset.order = order === 'asc' ? 'desc' : 'asc';
-                                link.querySelectorAll('.sort-icon').forEach(icon => {
-                                    icon.style.display = 'none';
-                                });
-                                link.querySelector(`.sort-icon.${order}`).style.display = 'inline';
-                            } else {
-                                link.dataset.order = 'desc';
-                                link.querySelectorAll('.sort-icon').forEach(icon => {
-                                    icon.style.display = 'none';
-                                });
-                            }
-                        });
-                    })
-                    .catch(error => {
-                        console.error('Error updating table:', error);
-                    });
-            }
-
-            // Xử lý sự kiện sắp xếp
-            sortLinks.forEach(link => {
-                link.addEventListener('click', function(event) {
-                    event.preventDefault();
-                    const column = link.dataset.column;
-                    const order = link.dataset.order;
-                    updateTable(currentQuery, currentProjectId, currentPage, column, order);
-                });
-            });
-
-            // Xử lý tìm kiếm
-            searchInput.addEventListener('input', function() {
-                updateTable(searchInput.value, currentProjectId, 1, sortColumn, sortOrder);
-            });
-
-            // Xử lý bộ lọc dự án
-            projectFilter.addEventListener('change', function() {
-                updateTable(searchInput.value, projectFilter.value, 1, sortColumn, sortOrder);
-            });
-
-            // Khởi tạo bảng với dữ liệu mặc định
-            updateTable(currentQuery, currentProjectId, currentPage, sortColumn, sortOrder);
-        });
-    </script>
+ 
 @endsection
